requirements.txt
Flask>=2.1.0
azure-identity>=1.14.0
azure-keyvault-secrets>=4.7.0
PyJWT>=2.6.0
gunicorn>=20.1.0
requests>=2.28.0
